split split_windows into three
